Make signatures consistent and pylint happy
