+++ tap_square/sync.py
@@ -1,4 +1,3 @@
-from datetime import timedelta
 import singer
 from singer import Transformer, metadata
 from .client import SquareClient
@@ -7,24 +6,13 @@ from .streams import STREAMS
 
 LOGGER = singer.get_logger()
 
-def get_date_windows(start_time):
-    window_start = singer.utils.strptime_to_utc(start_time)
-    now = singer.utils.now()
-    while window_start < now:
-        window_end = window_start + timedelta(days=7)
-        if window_end > now:
-            window_end = now
-        yield singer.utils.strftime(window_start), singer.utils.strftime(window_end)
-        window_start = window_end
-
 def sync(config, state, catalog): # pylint: disable=too-many-statements
     client = SquareClient(config)
 
     with Transformer() as transformer:
         for stream in catalog.get_selected_streams(state):
             tap_stream_id = stream.tap_stream_id
-            stream_obj = STREAMS[tap_stream_id](client, state)
-            replication_key = stream_obj.replication_key
+            stream_obj = STREAMS[tap_stream_id](client)
             stream_schema = stream.schema.to_dict()
             stream_metadata = metadata.to_map(stream.metadata)
 
@@ -40,7 +28,7 @@ def sync(config, state, catalog): # pylint: disable=too-many-statements
                 stream.replication_key
             )
 
-            state = stream_obj.sync(state, stream_schema, stream_metadata, config)
+            state = stream_obj.sync(state, stream_schema, stream_metadata, config, transformer)
             singer.write_state(state)
 
     state = singer.set_currently_syncing(state, None)
